+++ src/mesh/exodusII_io_helper.C
@@ -624,7 +624,7 @@ void ExodusII_IO_Helper::read_block_info()
 
 int ExodusII_IO_Helper::get_block_id(int index)
 {
-  libmesh_assert_less (static_cast<unsigned int>(index), block_ids.size());
+  libmesh_assert_less (index, block_ids.size());
 
   return block_ids[index];
 }
@@ -633,7 +633,7 @@ int ExodusII_IO_Helper::get_block_id(int index)
 
 std::string ExodusII_IO_Helper::get_block_name(int index)
 {
-  libmesh_assert_less (static_cast<unsigned int>(index), block_ids.size());
+  libmesh_assert_less (index, block_ids.size());
 
   return id_to_block_names[block_ids[index]];
 }
@@ -642,7 +642,7 @@ std::string ExodusII_IO_Helper::get_block_name(int index)
 
 int ExodusII_IO_Helper::get_side_set_id(int index)
 {
-  libmesh_assert_less (static_cast<unsigned int>(index), ss_ids.size());
+  libmesh_assert_less (index, ss_ids.size());
 
   return ss_ids[index];
 }
@@ -651,7 +651,7 @@ int ExodusII_IO_Helper::get_side_set_id(int index)
 
 std::string ExodusII_IO_Helper::get_side_set_name(int index)
 {
-  libmesh_assert_less (static_cast<unsigned int>(index), ss_ids.size());
+  libmesh_assert_less (index, ss_ids.size());
 
   return id_to_ss_names[ss_ids[index]];
 }
@@ -660,7 +660,7 @@ std::string ExodusII_IO_Helper::get_side_set_name(int index)
 
 int ExodusII_IO_Helper::get_node_set_id(int index)
 {
-  libmesh_assert_less (static_cast<unsigned int>(index), nodeset_ids.size());
+  libmesh_assert_less (index, nodeset_ids.size());
 
   return nodeset_ids[index];
 }
@@ -669,7 +669,7 @@ int ExodusII_IO_Helper::get_node_set_id(int index)
 
 std::string ExodusII_IO_Helper::get_node_set_name(int index)
 {
-  libmesh_assert_less (static_cast<unsigned int>(index), nodeset_ids.size());
+  libmesh_assert_less (index, nodeset_ids.size());
 
   return id_to_ns_names[nodeset_ids[index]];
 }
@@ -679,7 +679,7 @@ std::string ExodusII_IO_Helper::get_node_set_name(int index)
 
 void ExodusII_IO_Helper::read_elem_in_block(int block)
 {
-  libmesh_assert_less (static_cast<unsigned int>(block), block_ids.size());
+  libmesh_assert_less (block, block_ids.size());
 
   ex_err = exII::ex_get_elem_block(ex_id,
                                    block_ids[block],
@@ -803,11 +803,11 @@ void ExodusII_IO_Helper::read_nodeset_info()
 
 void ExodusII_IO_Helper::read_sideset(int id, int offset)
 {
-  libmesh_assert_less (static_cast<unsigned int>(id), ss_ids.size());
-  libmesh_assert_less (static_cast<unsigned int>(id), num_sides_per_set.size());
-  libmesh_assert_less (static_cast<unsigned int>(id), num_df_per_set.size());
-  libmesh_assert_less_equal (static_cast<unsigned int>(offset), elem_list.size());
-  libmesh_assert_less_equal (static_cast<unsigned int>(offset), side_list.size());
+  libmesh_assert_less (id, ss_ids.size());
+  libmesh_assert_less (id, num_sides_per_set.size());
+  libmesh_assert_less (id, num_df_per_set.size());
+  libmesh_assert_less_equal (offset, elem_list.size());
+  libmesh_assert_less_equal (offset, side_list.size());
 
   ex_err = exII::ex_get_side_set_param(ex_id,
                                        ss_ids[id],
@@ -846,9 +846,9 @@ void ExodusII_IO_Helper::read_sideset(int id, int offset)
 
 void ExodusII_IO_Helper::read_nodeset(int id)
 {
-  libmesh_assert_less (static_cast<unsigned int>(id), nodeset_ids.size());
-  libmesh_assert_less (static_cast<unsigned int>(id), num_nodes_per_set.size());
-  libmesh_assert_less (static_cast<unsigned int>(id), num_node_df_per_set.size());
+  libmesh_assert_less (id, nodeset_ids.size());
+  libmesh_assert_less (id, num_nodes_per_set.size());
+  libmesh_assert_less (id, num_node_df_per_set.size());
 
   ex_err = exII::ex_get_node_set_param(ex_id,
                                        nodeset_ids[id],
